chore(webrtc): update README and versioning, refactor WebRTC protocol handling
- Updated version in README.md to 4.0.6 and reflected changes in the VERSION file.
- Introduced a new ServerInfoProvider for handling version and filter queries.
- Refactored WebRTCManager to utilize the new ServerInfoProvider for control message handling.
- Removed the deprecated webrtc_protocol files and organized protocol-related functions into dedicated files.
- Enhanced the structure of the WebRTC protocol handling for improved maintainability and clarity.
