+++ src/cpp_accelerator/VERSION
@@ -1,1 +1,1 @@
-4.0.5
+4.0.6
