+++ src/cpp_accelerator/adapters/webrtc/BUILD
@@ -26,6 +26,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//proto:image_processor_service_cc_proto",
+        "//src/cpp_accelerator/adapters/webrtc/protocol:filter_resolver",
         "//src/cpp_accelerator/application/engine:engine",
         "@com_github_paullouisageneau_libdatachannel//:libdatachannel",
         "@spdlog",
@@ -38,19 +39,6 @@ cc_library(
 )
 
 cc_library(
-    name = "webrtc_protocol",
-    srcs = ["webrtc_protocol.cpp"],
-    hdrs = ["webrtc_protocol.h"],
-    deps = [
-        ":data_channel_framing",
-        "//proto:image_processor_service_cc_proto",
-        "//src/cpp_accelerator/application/engine:engine",
-        "@com_github_paullouisageneau_libdatachannel//:libdatachannel",
-        "@spdlog",
-    ],
-)
-
-cc_library(
     name = "webrtc_manager",
     srcs = ["webrtc_manager.cpp"],
     hdrs = ["webrtc_manager.h"],
@@ -58,12 +46,16 @@ cc_library(
     deps = [
         ":data_channel_framing",
         ":live_video_processor",
-        ":webrtc_protocol",
         "//proto:image_processor_service_cc_proto",
-        "//src/cpp_accelerator/application/engine:engine",
         "//src/cpp_accelerator/adapters/compute/cuda/memory:cuda_memory_pool",
+        "//src/cpp_accelerator/adapters/webrtc/protocol:filter_resolver",
+        "//src/cpp_accelerator/adapters/webrtc/protocol:message_codec",
+        "//src/cpp_accelerator/adapters/webrtc/protocol:session_routing",
+        "//src/cpp_accelerator/adapters/webrtc/sdp:sdp_utils",
+        "//src/cpp_accelerator/application/engine:engine",
+        "//src/cpp_accelerator/application/server_info:server_info_provider",
         "//src/cpp_accelerator/ports:i_media_session",
-        "@spdlog",
         "@com_github_paullouisageneau_libdatachannel//:libdatachannel",
+        "@spdlog",
     ],
 )
